requirements-dev.txt
django-nose==1.4.4
nose==1.3.7
model-mommy==1.3.2
